chore: 일지 작성 API 연결
