+++ src/api/index.ts
@@ -1,25 +1,10 @@
-import axios from 'axios';
+import { DayBook } from '@state/daybook';
+import client from './client';
+import { PostDaybookResponse } from './response';
 
-const client = axios.create({
-    baseURL: "http://10.10.6.131:8080",
-    headers: {
-        "Content-Type": "application/json"
-    },
-    withCredentials: true,
-});
-
-client.interceptors.request.use((config) => {
-    if (!config.headers) return config;
-
-    let token: string | null = null;
-
-    token = localStorage.getItem('accessToken');
-
-    if (token !== null) {
-      config.headers.Authorization = `Bearer ${token}`;
-    }
-
-    return config;
-})
-
-export default client;
+export const postDaybook = async (rawData: DayBook) => {
+  const { data } = await client.post<PostDaybookResponse>('/borad', {
+    rawData,
+  });
+  return data;
+};
