+++ src/api/response.ts
@@ -1,0 +1,9 @@
+export type PostDaybookResponse = {
+  boardId: number;
+  content: string;
+  authorId: number;
+  category: string;
+  hearts: number;
+  respectBoardId: number;
+  hashtags: string;
+};
